+++ src/main/java/com/ybe/tr1ll1on/domain/review/exception/ReviewAlreadyWrittenException.java
@@ -1,4 +1,4 @@
-package com.ybe.tr1ll1on.review.exception;
+package com.ybe.tr1ll1on.domain.review.exception;
 
 import com.ybe.tr1ll1on.global.exception.ExceptionCode;
 import com.ybe.tr1ll1on.global.exception.TrillionException;
