+++ src/main/java/com/ybe/tr1ll1on/domain/review/service/ReviewService.java
@@ -1,11 +1,12 @@
-package com.ybe.tr1ll1on.review.service;
+package com.ybe.tr1ll1on.domain.review.service;
 
 import com.ybe.tr1ll1on.domain.order.exception.OrderItemNotFoundException;
-import com.ybe.tr1ll1on.review.dto.request.ReviewCreateRequest;
-import com.ybe.tr1ll1on.review.dto.request.ReviewUpdateRequest;
-import com.ybe.tr1ll1on.review.dto.response.ReviewCreateResponse;
-import com.ybe.tr1ll1on.review.dto.response.ReviewListResponse;
-import com.ybe.tr1ll1on.review.exception.ReviewAlreadyWrittenException;
+import com.ybe.tr1ll1on.domain.review.dto.response.ReviewCreateResponse;
+import com.ybe.tr1ll1on.domain.review.exception.ReviewAlreadyWrittenException;
+import com.ybe.tr1ll1on.domain.review.model.Review;
+import com.ybe.tr1ll1on.domain.review.dto.request.ReviewCreateRequest;
+import com.ybe.tr1ll1on.domain.review.dto.request.ReviewUpdateRequest;
+import com.ybe.tr1ll1on.domain.review.dto.response.ReviewListResponse;
 import com.ybe.tr1ll1on.domain.order.model.OrderItem;
 import com.ybe.tr1ll1on.domain.order.repository.OrderItemRepository;
 import com.ybe.tr1ll1on.domain.product.model.Product;
@@ -13,9 +14,8 @@ import com.ybe.tr1ll1on.domain.product.model.Product;
 import com.ybe.tr1ll1on.domain.user.model.User;
 import com.ybe.tr1ll1on.domain.user.repository.UserRepository;
 import com.ybe.tr1ll1on.global.exception.TrillionExceptionCode;
-import com.ybe.tr1ll1on.review.exception.ReviewNotFoundException;
-import com.ybe.tr1ll1on.review.model.Review;
-import com.ybe.tr1ll1on.review.repository.ReviewRepository;
+import com.ybe.tr1ll1on.domain.review.exception.ReviewNotFoundException;
+import com.ybe.tr1ll1on.domain.review.repository.ReviewRepository;
 import com.ybe.tr1ll1on.security.util.SecurityUtil;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
